+++ booksCrudService/src/main/java/com/books/crud/booksdCrud/livro/LivroAtualizarDTO.java
@@ -1,12 +1,14 @@
 package com.books.crud.booksdCrud.livro;
 
+import java.math.BigDecimal;
+
 public record LivroAtualizarDTO(
 
     String nome,
 
     String capa,
 
-    Double valor,
+    BigDecimal valor,
 
     String descricao
 ) {
