+++ booksCrudService/src/main/java/com/books/crud/booksdCrud/livro/LivroResponseDTO.java
@@ -1,11 +1,13 @@
 package com.books.crud.booksdCrud.livro;
 
+import java.math.BigDecimal;
+
 public record LivroResponseDTO(
     Long id,
     String nome,
     Boolean ativo,
     String capa,
-    Double valor,
+    BigDecimal valor,
     String descricao,
     Integer estoque
 ) {
